Fix coding standards (phpcs) errors, phpunit errors and risky tests
