+++ classes/user_filter_activeonldap.php
@@ -1,0 +1,46 @@
+<?php
+// This file is part of Moodle - http://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
+
+/**
+ * This file contains the User Filter API for activeonldap.
+ *
+ * @package   tool_ldapsync
+ * @copyright 2019 onwards, The Regents of the University of California
+ * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+namespace tool_ldapsync;
+
+defined('MOODLE_INTERNAL') || die();
+
+/**
+ * Generic yes/no filter with radio buttons for integer fields.
+ */
+class user_filter_activeonldap extends \user_filter_yesno {
+    /**
+     * Returns the condition to be used with SQL
+     *
+     * @param array $data filter settings
+     * @return array sql string and $params
+     */
+    public function get_sql_filter($data) {
+
+        $value = $data['value'];
+        $not = $value ? '' : 'NOT';
+
+        return ["username $not IN ( SELECT cn FROM {tool_ldapsync} )", []];
+    }
+}
