+++ tests/plugin_test.php
@@ -27,35 +27,13 @@
  * define('TEST_AUTH_LDAP_DOMAIN', 'dc=example,dc=local');
  *
  * @package    tool_ldapsync
- * @copyright  Copyright (c) 2019, UCSF Center for Knowledge Management
- * @author     2019 Carson Tam {@email carson.tam@ucsf.edu}
+ * @copyright  2019 onwards, The Regents of the University of California
+ * @author     Carson Tam {@email carson.tam@ucsf.edu}
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
 defined('MOODLE_INTERNAL') || die();
-
-/**
- * Testable object for the importer
- */
-class Testable_tool_ldapsync_importer_for_plugin extends \tool_ldapsync\importer {
-    /**
-     * Change the visibility scope of the protected function to public
-     */
-    public function connecttoldap() {
-        return parent::connecttoldap();
-    }
-
-    /**
-     * Searches LDAP for user records that were updated/created after a given datetime.
-     * @param \LDAP\Connection $ldap the LDAP connection
-     * @param string|null $ldaptimestamp the datetime
-     * @return array nested array of user records
-     * @throws Exception if search fails
-     */
-    public function getupdatesfromldap($ldap, $ldaptimestamp = null) {
-        return parent::getupdatesfromldap($ldap, $ldaptimestamp);
-    }
-}
+require_once(__DIR__ . '/classes/testable_tool_ldapsync_importer_for_plugin.php');
 
 /**
  * Test case for ldapsync plugin
@@ -107,19 +85,19 @@ final class plugin_test extends advanced_testcase {
             $this->markTestSkipped('Can not connect to LDAP test server: ' . $debuginfo);
         }
 
-        $this->ldapConn = $connection;
+        $this->ldapconn = $connection;
 
         // Create new empty test container.
         $topdn = 'dc=moodletest,' . TEST_TOOL_LDAPSYNC_DOMAIN;
 
         // Let tearDown() handle it.
-        $this->recursive_delete($this->ldapConn, TEST_TOOL_LDAPSYNC_DOMAIN, 'dc=moodletest');
+        $this->recursive_delete($this->ldapconn, TEST_TOOL_LDAPSYNC_DOMAIN, 'dc=moodletest');
 
         $o = [];
         $o['objectClass'] = ['dcObject', 'organizationalUnit'];
         $o['dc']         = 'moodletest';
         $o['ou']         = 'MOODLETEST';
-        if (!ldap_add($this->ldapConn, 'dc=moodletest,' . TEST_AUTH_LDAP_DOMAIN, $o)) {
+        if (!ldap_add($this->ldapconn, 'dc=moodletest,' . TEST_AUTH_LDAP_DOMAIN, $o)) {
             $this->markTestSkipped('Can not create test LDAP container.');
         }
 
@@ -127,7 +105,7 @@ final class plugin_test extends advanced_testcase {
         $o = [];
         $o['objectClass'] = ['organizationalUnit'];
         $o['ou']          = 'users';
-        ldap_add($this->ldapConn, 'ou=' . $o['ou'] . ',' . $topdn, $o);
+        ldap_add($this->ldapconn, 'ou=' . $o['ou'] . ',' . $topdn, $o);
 
         $gmtts = strtotime('2000-01-01 00:00:00');
 
@@ -174,7 +152,7 @@ final class plugin_test extends advanced_testcase {
         set_config('field_updateremote_idnumber', '0', 'auth_tool_ldapsync');
         set_config('field_lock_idnumber', 'unlocked', 'auth_tool_ldapsync');
 
-        $this->sync = new Testable_tool_ldapsync_importer_for_plugin($gmtts);
+        $this->sync = new testable_tool_ldapsync_importer_for_plugin($gmtts);
 
         ob_start();
     }
@@ -183,14 +161,14 @@ final class plugin_test extends advanced_testcase {
      * Tear down test case
      */
     protected function tearDown(): void {
-        if (!$this->ldapConn) {
-            $this->recursive_delete($this->ldapConn, TEST_TOOL_LDAPSYNC_DOMAIN, 'dc=moodletest');
-            ldap_close($this->ldapConn);
-            $this->ldapConn = null;
-        }
+        if ($this->ldapconn) {
+            $this->recursive_delete($this->ldapconn, TEST_TOOL_LDAPSYNC_DOMAIN, 'dc=moodletest');
+            ldap_close($this->ldapconn);
+            $this->ldapconn = null;
 
-        // Use ob_end_flush() to flush to standard output.
-        ob_end_clean();
+            // Use ob_end_flush() to flush to standard output.
+            ob_end_clean();
+        }
         parent::tearDown();
     }
 
@@ -221,7 +199,7 @@ final class plugin_test extends advanced_testcase {
         // Create a few users
         $topdn = 'dc=moodletest,' . TEST_TOOL_LDAPSYNC_DOMAIN;
         for ($i = 1; $i <= 5; $i++) {
-            $this->create_ldap_user($this->ldapConn, $topdn, $i);
+            $this->create_ldap_user($this->ldapconn, $topdn, $i);
         }
 
         // Test with current time + 7 days, expect no update returned
@@ -267,7 +245,7 @@ final class plugin_test extends advanced_testcase {
 
         // Create a few users.
         for ($i = 1; $i <= 5; $i++) {
-            $this->create_ldap_user($this->ldapConn, $topdn, $i);
+            $this->create_ldap_user($this->ldapconn, $topdn, $i);
         }
 
         $this->assertEquals(2, $DB->count_records('user'));
