+++ tests/purgeusers_test.php
@@ -27,27 +27,14 @@
  * define('TEST_AUTH_LDAP_DOMAIN', 'dc=example,dc=local');
  *
  * @package    tool_ldapsync
- * @copyright  Copyright (c) 2020, UCSF Center for Knowledge Management
- * @author     2020 Carson Tam {@email carson.tam@ucsf.edu}
+ * @copyright  2019 onwards, The Regents of the University of California
+ * @author     Carson Tam {@email carson.tam@ucsf.edu}
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
 defined('MOODLE_INTERNAL') || die();
+require_once(__DIR__ . '/classes/testable_tool_ldapsync_importer_for_purgeusers.php');
 
-/**
- * Testable object for the importer
- */
-class Testable_tool_ldapsync_importer_for_purgeusers extends \tool_ldapsync\importer {
-    /**
-     * Get updates from LDAP
-     * @param connection $ldap
-     * @param string $ldaptimestamp
-     */
-    public function getupdatesfromldap($ldap, $ldaptimestamp = null) {
-        // Change visibility to allow tests to call protected function.
-        return parent::getupdatesfromldap($ldap, $ldaptimestamp);
-    }
-}
 /**
  * Test case for purgeusers
  */
@@ -97,19 +84,19 @@ final class purgeusers_test extends advanced_testcase {
             $this->markTestSkipped('Can not connect to LDAP test server: ' . $debuginfo);
         }
 
-        $this->ldapConn = $connection;
+        $this->ldapconn = $connection;
 
         // Create new empty test container.
         $topdn = 'dc=moodletest,' . TEST_TOOL_LDAPSYNC_DOMAIN;
 
         // Let tearDown() handle it.
-        $this->recursive_delete($this->ldapConn, TEST_TOOL_LDAPSYNC_DOMAIN, 'dc=moodletest');
+        $this->recursive_delete($this->ldapconn, TEST_TOOL_LDAPSYNC_DOMAIN, 'dc=moodletest');
 
         $o = [];
         $o['objectClass'] = ['dcObject', 'organizationalUnit'];
         $o['dc']         = 'moodletest';
         $o['ou']         = 'MOODLETEST';
-        if (!ldap_add($this->ldapConn, 'dc=moodletest,' . TEST_AUTH_LDAP_DOMAIN, $o)) {
+        if (!ldap_add($this->ldapconn, 'dc=moodletest,' . TEST_AUTH_LDAP_DOMAIN, $o)) {
             $this->markTestSkipped('Can not create test LDAP container.');
         }
 
@@ -117,7 +104,7 @@ final class purgeusers_test extends advanced_testcase {
         $o = [];
         $o['objectClass'] = ['organizationalUnit'];
         $o['ou']          = 'users';
-        ldap_add($this->ldapConn, 'ou=' . $o['ou'] . ',' . $topdn, $o);
+        ldap_add($this->ldapconn, 'ou=' . $o['ou'] . ',' . $topdn, $o);
 
         $gmtts = strtotime('2000-01-01 00:00:00');
 
@@ -153,7 +140,7 @@ final class purgeusers_test extends advanced_testcase {
         set_config('field_updateremote_idnumber', '0', 'tool_ldapsync');
         set_config('field_lock_idnumber', 'unlocked', 'tool_ldapsync');
 
-        $this->sync = new Testable_tool_ldapsync_importer_for_purgeusers($gmtts);
+        $this->sync = new testable_tool_ldapsync_importer_for_purgeusers($gmtts);
 
         ob_start();
     }
@@ -162,14 +149,13 @@ final class purgeusers_test extends advanced_testcase {
      * Tear down test case
      */
     protected function tearDown(): void {
-        if (!$this->ldapConn) {
-            $this->recursive_delete($this->ldapConn, TEST_TOOL_LDAPSYNC_DOMAIN, 'dc=moodletest');
-            ldap_close($this->ldapConn);
-            $this->ldapConn = null;
+        if ($this->ldapconn) {
+            $this->recursive_delete($this->ldapconn, TEST_TOOL_LDAPSYNC_DOMAIN, 'dc=moodletest');
+            ldap_close($this->ldapconn);
+            $this->ldapconn = null;
+            // Use ob_end_flush() if you want to see the output.
+            ob_end_clean();
         }
-
-        // Use ob_end_flush() if you want to see the output.
-        ob_end_clean();
         parent::tearDown();
     }
 
@@ -187,7 +173,7 @@ final class purgeusers_test extends advanced_testcase {
         // Create a few users.
         $topdn = 'dc=moodletest,' . TEST_TOOL_LDAPSYNC_DOMAIN;
         for ($i = 1; $i <= 5; $i++) {
-            $this->create_ldap_user($this->ldapConn, $topdn, $i);
+            $this->create_ldap_user($this->ldapconn, $topdn, $i);
             $this->assertTrue($this->sync->check_users_in_ldap('00000' . $i . '@ucsf.edu'));
         }
